# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MobiDuka Retail POS API Blueprint",
    "description": "Multi-tenant, offline-first production backend engine supporting Flutter mobile clients and Kenyan retail ecosystems (Cash tracking, M-Pesa records, local synchronization syncing).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current app origin API router"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /auth/login. Use the token as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Multi-Factor Authentication Entry Gateway",
        "description": "Handles standard credential authorization maps for owners and fast PIN code verification sequences for shift cashiers.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Use the owner password example or the cashier PIN example. Do not send password and pin together.",
                "properties": {
                  "identifier": { "type": "string" },
                  "password": { "type": "string" },
                  "pin": { "type": "string" }
                }
              },
              "examples": {
                "ownerPassword": {
                  "summary": "Seeded owner password login",
                  "value": { "identifier": "owner@mobiduka.com", "password": "OwnerPass123" }
                },
                "cashierPin": {
                  "summary": "Seeded cashier PIN login",
                  "value": { "identifier": "cashier1", "pin": "2540" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Authentication successful, returns active JWT token parameters." },
          "401": { "description": "Invalid user credentials or mismatched secret keys." }
        }
      }
    },
...
    "/suppliers": {
      "get": { "tags": ["Business"], "summary": "List suppliers for a business", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }], "responses": { "200": { "description": "Suppliers returned with product and purchase-order counts." }, "401": { "description": "Authenticated business context is required." } } },
      "post": { "tags": ["Business"], "summary": "Create a supplier", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["businessId", "name"], "properties": { "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "name": { "type": "string" }, "category": { "type": "string" }, "contactPerson": { "type": "string" }, "phone": { "type": "string" }, "email": { "type": "string" }, "location": { "type": "string" }, "notes": { "type": "string" } } } } } }, "responses": { "201": { "description": "Supplier created." }, "400": { "description": "Invalid supplier data." }, "409": { "description": "Supplier already exists." } } },
      "patch": { "tags": ["Business"], "summary": "Update a supplier", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["businessId", "id"], "properties": { "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }, "category": { "type": "string" }, "contactPerson": { "type": "string" }, "phone": { "type": "string" }, "email": { "type": "string" }, "location": { "type": "string" }, "notes": { "type": "string" } } } } } }, "responses": { "200": { "description": "Supplier updated." }, "404": { "description": "Supplier not found." }, "409": { "description": "Supplier name already exists." } } },
      "delete": { "tags": ["Business"], "summary": "Delete an unused supplier", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }, { "name": "id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Supplier deleted." }, "404": { "description": "Supplier not found." }, "409": { "description": "Supplier is referenced by products or purchase orders." } } }
    },
...
    "/products": {
      "get": { "tags": ["Inventory"], "summary": "List products and inventory", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }], "responses": { "200": { "description": "Product catalog returned." }, "401": { "description": "Authenticated business context is required." } } },
      "post": { "tags": ["Inventory"], "summary": "Create a product", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "name": { "type": "string" }, "sellingPrice": { "type": "number" }, "costPrice": { "type": "number" }, "sku": { "type": "string" }, "barcode": { "type": "string" }, "categoryId": { "type": "string", "format": "uuid" }, "emoji": { "type": "string" } } } } } }, "responses": { "201": { "description": "Product created." }, "401": { "description": "Authenticated business context is required." } } },
      "patch": { "tags": ["Inventory"], "summary": "Update a product and optionally set stock", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["id", "businessId"], "properties": { "id": { "type": "string", "format": "uuid" }, "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "name": { "type": "string" }, "sellingPrice": { "type": "number" }, "costPrice": { "type": "number" }, "minimumStock": { "type": "integer", "minimum": 0 }, "stock": { "type": "integer", "minimum": 0 }, "categoryId": { "type": "string", "format": "uuid" }, "emoji": { "type": "string" } } } } } }, "responses": { "200": { "description": "Product updated." }, "400": { "description": "Invalid product data." }, "404": { "description": "Product not found." } } },
      "delete": { "tags": ["Inventory"], "summary": "Soft-delete a product or hard-delete an unreferenced product", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }, { "name": "id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "hard", "in": "query", "required": false, "schema": { "type": "boolean", "default": false } }], "responses": { "200": { "description": "Product deleted." }, "400": { "description": "Required identifiers are missing." }, "404": { "description": "Product not found." }, "409": { "description": "Product has historical records and can only be soft-deleted." } } }
    },
...
  }
}